General UI update
This update makes sure the user provides valid yes/no responses to inputs on whether they have a static IP, a netmark, a gateway, DNS server names as well as if they want to reset the file from the backup of the interfaces file. The code also outputs information on the changes made before the code exits.

More work need to be done, particularly testing (Tyler is running the D3S and Raspberry PI 3 right now!) and to ensure proper dependence/independence of the netmask, gateway, and DNS server names based on further Internet research.